fix(conflicts,pulls): flag only live resolve streams as interrupted
- `ConflictResolveView` now tracks an explicit `runningRef` instead of deriving in-flight state from `phase`/`generating`. Both lag the decision points: `generating` stays true across a cancel's kill round-trip and the phase is still "streaming" while a finished proposal is diffed, so an `<Activity>` hide right after a cancel or a completed run could mark the session interrupted and have the kickoff restart it. The flag is cleared on error, on an AI-ignored file, on completion, and on user cancel, so teardown only resumes a stream it actually cut short.
- `ChecksRollup` decides `needsRequiredAttention` once per render into a Map keyed by the check object (names are not unique). The attention count, the sort comparator, and every row read that flag rather than each re-running a scan over the whole check list.
- `isOutstanding` moves from `useBranchRequiredChecks` into `check-presentation`, so the rollup no longer imports from the hook and both surfaces read one marker set from the pure presentation module.
